requirements.txt
Flask>=2.0.0
requests>=2.25.0
Pillow>=9.0.0
python-dotenv>=0.19.0
